Remove tags.go
